requirements.txt
google-adk
google-genai
fastapi
uvicorn
python-dotenv
